requirements.txt
google-api-python-client
textblob
matplotlib
urllib3
streamlit
python-dotenv
requests
matplotlib